Запуск job-ов deployment и app * bootstrap перезапустится, если предыдущий запуск завершился с ошибкой. * before_apply_job перезапускается при каждом apply.
