+++ lib/dapp.rb
@@ -71,6 +71,7 @@ require 'dapp/deployment/config/config'
 require 'dapp/deployment/core_ext/hash'
 require 'dapp/deployment/mod/namespace'
 require 'dapp/deployment/mod/system_environments'
+require 'dapp/deployment/mod/jobs'
 require 'dapp/deployment/app'
 require 'dapp/deployment/deployment'
 require 'dapp/deployment/kube_base'
